// src/types.ts
export interface Translation {
  nav: {
    home: string;
    about: string;
    skills: string;
    projects: string;
    contact: string;
  };
  hero: {
    greeting: string;
    title: string;
    subtitle: string;
  };
  about: {
    title: string;
    description: string;
  };
  skills: {
    title: string;
    business: string;
    marketing: string;
    design: string;
  };
  contact: {
    title: string;
    getInTouch: string;
    email: string;
  };
}